+++ src/ScottPlot4/ScottPlot.Eto/PlotView.cs
@@ -5,7 +5,7 @@ using Eto.Drawing;
 
 namespace ScottPlot.Eto
 {
-    public class PlotView : ImageView // ported from ScottPlot..WpPlot.axaml.cs v4.1.27
+    public class PlotView : ImageView, ScottPlot.Control.IPlotControl // ported from ScottPlot..WpPlot.axaml.cs v4.1.27
     {
         /// <summary>
         /// This is the plot displayed by the user control.
@@ -16,7 +16,7 @@ namespace ScottPlot.Eto
         /// <summary>
         /// This object can be used to modify advanced behaior and customization of this user control.
         /// </summary>
-        public readonly Control.Configuration Configuration;
+        public Control.Configuration Configuration { get; }
 
         /// <summary>
         /// This event is invoked any time the axis limits are modified.
@@ -115,6 +115,11 @@ namespace ScottPlot.Eto
         /// Reset this control by replacing the current plot with an existing plot
         /// </summary>
         public void Reset(Plot newPlot) => Backend.Reset(Width, Height, newPlot);
+
+        /// <summary>
+        /// Re-render the plot and update the image displayed by this control.
+        /// </summary>
+        public void Refresh() => Refresh(false);
 
         /// <summary>
         /// Re-render the plot and update the image displayed by this control.