+++ src/ScottPlot4/autoformat.sh
@@ -4,4 +4,4 @@
 dotnet tool update -g dotnet-format
 
 # Run the dotnet formatter
-dotnet format
+dotnet format ScottPlot4.sln