+++ CHANGELOG.md
@@ -30,6 +30,7 @@ _not yet published on NuGet..._
 * Heatmap: Added support for transparent single-color heatmaps (#2336) _Thanks @bukkideme_
 * Statistics: Improved median calculation method in population plots (#2363) _Thanks @Syntaxrabbit_
 * AxisLineVector: Improved automatic axis limits when using limited axis lines (#2371) _Thanks @ChrisAtVault_
+* Controls: Configuration.AddLinkedControl() simplifies axis sharing across multiple controls (#2402, #2372)
 
 ## ScottPlot 4.1.60
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2022-12-23_